feat(programming-1): split train/validate/test data

+++ programming/assignment1/interactive_naive_bayes/naive_bayes/classifier.py
@@ -6,7 +6,7 @@ import numpy.typing as npt
 import scipy.special
 
 Category = np.int_
-Count = np.uint
+Count = np.int_
 
 
 @dataclasses.dataclass
